mubs_241/lab/src/requirements.txt
matplotlib==3.6.2
numpy==1.23.5
PyYAML==6.0
scipy==1.9.3
